+++ catalog/app/containers/Bucket/PackageDirectoryDialog.tsx
@@ -12,6 +12,8 @@ import * as AWS from 'utils/AWS'
 import * as Data from 'utils/Data'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import StyledLink from 'utils/StyledLink'
+import * as s3paths from 'utils/s3paths'
+import * as packageHandle from 'utils/packageHandle'
 import * as validators from 'utils/validators'
 import type * as workflows from 'utils/workflows'
 
@@ -200,24 +202,52 @@ function DialogForm({
     [editorElement, handleNameChange, setMetaHeight],
   )
 
+  const username = redux.useSelector(authSelectors.username)
+
+  const getDefaultPackageName = React.useCallback(
+    (workflow) =>
+      packageHandle.convert(workflow?.packageHandle, 'files', {
+        directory: s3paths.ensureNoSlash(path),
+        username: PD.getUsernamePrefix(username),
+      }),
+    [path, username],
+  )
+
+  const [initialValues, setInitialValues] = React.useState({
+    files: initialFiles,
+    meta: PD.EMPTY_META_VALUE,
+    name: getDefaultPackageName(selectedWorkflow || workflowsConfig),
+    workflow: selectedWorkflow,
+  })
+
+  const onWorkflowChange = React.useCallback(
+    ({ modified, values }) => {
+      setWorkflow(values.workflow)
+
+      if (modified.name) return
+      const defaultPackageName = getDefaultPackageName(values.workflow)
+      setInitialValues(R.assoc('name', defaultPackageName, values))
+    },
+    [getDefaultPackageName, setWorkflow],
+  )
+
   React.useEffect(() => {
     if (document.body.contains(editorElement)) {
       setMetaHeight(editorElement!.clientHeight)
     }
   }, [editorElement, setMetaHeight])
 
-  const username = redux.useSelector(authSelectors.username)
-  const usernamePrefix = React.useMemo(() => PD.getUsernamePrefix(username), [username])
-
   return (
     <RF.Form
+      initialValues={initialValues}
       onSubmit={onSubmitWrapped}
       subscription={{
-        submitting: true,
-        submitFailed: true,
         error: true,
-        submitError: true,
         hasValidationErrors: true,
+        initialValues: true,
+        submitError: true,
+        submitFailed: true,
+        submitting: true,
       }}
     >
       {({
@@ -244,7 +274,7 @@ function DialogForm({
                 subscription={{ modified: true, values: true }}
                 onChange={({ modified, values }) => {
                   if (modified?.workflow && values.workflow !== selectedWorkflow) {
-                    setWorkflow(values.workflow)
+                    onWorkflowChange({ modified, values })
                   }
                 }}
               />
@@ -259,7 +289,6 @@ function DialogForm({
                     component={PD.WorkflowInput}
                     name="workflow"
                     workflowsConfig={workflowsConfig}
-                    initialValue={selectedWorkflow}
                     validate={
                       validators.required as FF.FieldValidator<workflows.Workflow>
                     }
@@ -271,7 +300,6 @@ function DialogForm({
 
                   <RF.Field
                     component={PD.PackageNameInput}
-                    initialValue={usernamePrefix}
                     name="name"
                     validate={validators.composeAsync(
                       validators.required,
@@ -311,7 +339,6 @@ function DialogForm({
                       validate={validateMetaInput}
                       validateFields={['meta']}
                       isEqual={R.equals}
-                      initialValue={PD.EMPTY_META_VALUE}
                       ref={setEditorElement}
                     />
                   )}
@@ -332,7 +359,6 @@ function DialogForm({
                     }}
                     title="Select files and directories to package"
                     isEqual={R.equals}
-                    initialValue={initialFiles}
                     truncated={truncated}
                   />
                 </PD.RightColumn>
@@ -457,6 +483,7 @@ export default function PackageDirectoryDialog({
     })
     if (onClose) onClose()
     setSuccess(null)
+    setWorkflow(undefined)
   }, [submitting, success, setSuccess, onClose, onExited])
 
   const handleExited = React.useCallback(() => {